requirements.txt
pyautogui>=0.9.54
Pillow>=10.4.0
opencv-python>=4.10.0.84
pytesseract>=0.3.13
easyocr>=1.7.0
boto3>=1.34.71
PySide6>=6.8.0
matplotlib>=3.8.4
numpy>=1.26.4
pandas>=2.2.0
PyYAML>=6.0.2
pygetwindow>=0.0.9
